must be friends to transfer,some more checks

+++ src/main/java/com/techelevator/tenmo/controller/AccountController.java
@@ -1,12 +1,14 @@
 package com.techelevator.tenmo.controller;
 
 import com.techelevator.tenmo.dao.AccountDao;
+import com.techelevator.tenmo.dao.UserDao;
 import com.techelevator.tenmo.model.Account;
 import com.techelevator.tenmo.model.AccountDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.security.Principal;
@@ -17,6 +19,8 @@ import java.util.List;
 public class AccountController {
     @Autowired
     AccountDao jdbcAccount;
+    @Autowired
+    UserDao jdbcUser;
 
 
     @RequestMapping(path = "/account", method = RequestMethod.GET)
@@ -30,13 +34,23 @@ public class AccountController {
     }
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @RequestMapping(path = "/account", method = RequestMethod.DELETE)
-    public void delete(@Valid @RequestBody Account account){
-        jdbcAccount.deleteAccount(account);
+    public void delete(@Valid @RequestBody Account account, Principal principal){
+        int userId = jdbcUser.findIdByUsername(principal.getName());
+        if(userId != account.getUserId()){
+            throw new ResponseStatusException(HttpStatus.EXPECTATION_FAILED, "Not Authorized. Please sign in as the correct user.");
+        }else {
+            jdbcAccount.deleteAccount(account);
+        }
     }
 
     @RequestMapping(path = "/account", method = RequestMethod.PUT)
-    public Account update(@Valid @RequestBody Account account){
-        return jdbcAccount.updateAccount(account);
+    public Account update(@Valid @RequestBody Account account, Principal principal){
+        int userId = jdbcUser.findIdByUsername(principal.getName());
+        if(userId != account.getUserId()){
+            throw new ResponseStatusException(HttpStatus.EXPECTATION_FAILED, "Not Authorized. Please sign in as the correct user.");
+        }else {
+            return jdbcAccount.updateAccount(account);
+        }
     }
 
     @RequestMapping(path = "/account/{accountId}", method = RequestMethod.GET)
